Split out update-email from user dashboard
